+++ nbproject/private/private.properties
@@ -3,4 +3,4 @@ auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
 endorsed.javafx.ant.classpath=.
 javafx.run.inbrowser=<Default System Browser>
 javafx.run.inbrowser.path=C:\\Program Files\\Internet Explorer\\iexplore.exe
-user.properties.file=C:\\Users\\user\\AppData\\Roaming\\NetBeans\\16\\build.properties
+user.properties.file=C:\\Users\\Blizzard\\AppData\\Roaming\\NetBeans\\19\\build.properties
