+++ src/multiformjfx1/Controller/Ubah_Harga/FXML_ubahhargaController.java
@@ -2,8 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/javafx/FXMLController.java to edit this template
  */
-package multiformjfx1;
+package multiformjfx1.Controller.Ubah_Harga;
 
+import multiformjfx1.Controller.Input.FXMLDocumentController;
 import java.net.URL;
 import java.util.ResourceBundle;
 import javafx.event.ActionEvent;
@@ -18,9 +19,7 @@ import javafx.scene.control.TextField;
  * @author Ramdani
  */
 public class FXML_ubahhargaController implements Initializable {
-    
-    
-    
+
     @FXML
     private TextField txthrsederhana;
     @FXML
@@ -44,17 +43,16 @@ public class FXML_ubahhargaController implements Initializable {
     @Override
     public void initialize(URL url, ResourceBundle rb) {
         // TODO
-    }    
+    }
 
     @FXML
     private void okklik(ActionEvent event) {
-        FXMLDocumentController.dtharga.setHrsederhana(Integer.parseInt
-        (txthrsederhana.getText()));       FXMLDocumentController.dtharga.setHrmenengah(Integer.parseInt
-        (txthrmenengah.getText()));FXMLDocumentController.dtharga.setHrmewah(Integer.parseInt
-        (txthrmewah.getText()));FXMLDocumentController.dtharga.setHtsederhana(Integer.parseInt
-        (txthtsederhana.getText()));FXMLDocumentController.dtharga.setHtmenengah(Integer.parseInt
-        (txthtmenengah.getText())); FXMLDocumentController.dtharga.setHtmewah(Integer.parseInt
-        (txthtmewah.getText()));
+        FXMLDocumentController.dtharga.setHrsederhana(Integer.parseInt(txthrsederhana.getText()));
+        FXMLDocumentController.dtharga.setHrmenengah(Integer.parseInt(txthrmenengah.getText()));
+        FXMLDocumentController.dtharga.setHrmewah(Integer.parseInt(txthrmewah.getText()));
+        FXMLDocumentController.dtharga.setHtsederhana(Integer.parseInt(txthtsederhana.getText()));
+        FXMLDocumentController.dtharga.setHtmenengah(Integer.parseInt(txthtmenengah.getText()));
+        FXMLDocumentController.dtharga.setHtmewah(Integer.parseInt(txthtmewah.getText()));
         btnok.getScene().getWindow().hide();
 
     }
@@ -66,8 +64,8 @@ public class FXML_ubahhargaController implements Initializable {
         txthrmewah.setText("500000000");
         txthtsederhana.setText("500000");
         txthtmenengah.setText("800000");
-        txthtmewah.setText("2000000"); 
+        txthtmewah.setText("2000000");
 
     }
-    
+
 }
