+++ src/multiformjfx1/FXML/Output/FXML_output.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.*?>
 
 
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="501.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" fx:controller="multiformjfx1.FXML_outputController">
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="501.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" fx:controller="multiformjfx1.Controller.Output.FXML_outputController">
    <children>
       <Label layoutX="52.0" layoutY="37.0" text="Nama Pembeli" />
       <Label fx:id="lblpembeli" layoutX="372.0" layoutY="37.0" text="Label" />
