+++ src/multiformjfx1/Model/Harga/Model_harga.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package multiformjfx1;
+package multiformjfx1.Model.Harga;
 
 
 public class Model_harga {
